refactor: moved queryTerrainElevation to helpers as queryElevationFromRasterDEM
