+++ src/lib/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './capitalize';
 export * from './cleanMaplibreStyle';
 export * from './debounce';
+export * from './queryElevationFromRasterDEM';
