+++ src/lib/interfaces/index.ts
@@ -1,8 +1,9 @@
-export * from './TerradrawControlOptions';
+export * from './AreaUnit';
+export * from './DistanceUnit';
 export * from './EventTypes';
 export * from './MeasureControlOptions';
 export * from './ModeOptions';
+export * from './TerrainSource';
+export * from './TerradrawControlOptions';
 export * from './TerradrawMode';
 export * from './TerradrawModeClass';
-export * from './AreaUnit';
-export * from './DistanceUnit';
